# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SwitchPoint API",
    "description": "Implements the FCA-prescribed pension switching, DB transfer and cashflow analyses for UK IFAs. Authenticate with POST /api/v1/auth/login and send 'Authorization: Bearer <token>'.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5093/"
    }
  ],
  "paths": {
...
    "/api/v1/calculations/pension-switch": {
      "post": {
        "tags": [
          "Calculations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchCalcRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchCalcRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchCalcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PensionSwitchResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calculations/db-transfer": {
      "post": {
        "tags": [
          "Calculations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferCalcRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferCalcRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferCalcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbTransferResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calculations/cashflow": {
      "post": {
        "tags": [
          "Calculations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowCalcRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowCalcRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowCalcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calculations/cashflow/stochastic": {
      "post": {
        "tags": [
          "Calculations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowCalcRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowCalcRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowCalcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StochasticResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calculations/riy": {
      "post": {
        "tags": [
          "Calculations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiyCalcRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RiyCalcRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RiyCalcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calculations/tax": {
      "post": {
        "tags": [
          "Calculations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxCalcRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxCalcRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TaxCalcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxComputationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/pension-switch": {
      "post": {
        "tags": [
          "PensionSwitchAnalyses"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchAnalysisWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchAnalysisWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchAnalysisWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PensionSwitchAnalysisDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/pension-switch/{id}": {
      "get": {
        "tags": [
          "PensionSwitchAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PensionSwitchAnalysisDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PensionSwitchAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchAnalysisWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchAnalysisWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PensionSwitchAnalysisWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PensionSwitchAnalysisDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PensionSwitchAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/v1/analyses/pension-switch/{id}/calculate": {
      "post": {
        "tags": [
          "PensionSwitchAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PensionSwitchAnalysisDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/pension-switch/{id}/lock": {
      "post": {
        "tags": [
          "PensionSwitchAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PensionSwitchAnalysisDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/db-transfer": {
      "post": {
        "tags": [
          "DbTransferAnalyses"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferAnalysisWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferAnalysisWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferAnalysisWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbTransferAnalysisDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/db-transfer/{id}": {
      "get": {
        "tags": [
          "DbTransferAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbTransferAnalysisDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "DbTransferAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferAnalysisWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferAnalysisWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DbTransferAnalysisWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbTransferAnalysisDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DbTransferAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/v1/analyses/db-transfer/{id}/calculate": {
      "post": {
        "tags": [
          "DbTransferAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbTransferAnalysisDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/db-transfer/{id}/lock": {
      "post": {
        "tags": [
          "DbTransferAnalyses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbTransferAnalysisDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/cashflow": {
      "post": {
        "tags": [
          "CashflowPlans"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowPlanWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowPlanWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowPlanWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowPlanDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/cashflow/{id}": {
      "get": {
        "tags": [
          "CashflowPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowPlanDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CashflowPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowPlanWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowPlanWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowPlanWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowPlanDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CashflowPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/v1/analyses/cashflow/{id}/calculate": {
      "post": {
        "tags": [
          "CashflowPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowPlanDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/cashflow/{id}/calculate/stochastic": {
      "post": {
        "tags": [
          "CashflowPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowPlanDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/cashflow/{id}/lock": {
      "post": {
        "tags": [
          "CashflowPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowPlanDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CedingSchemeRef": {
        "required": [
          "schemeId",
          "inline"
        ],
        "type": "object",
        "properties": {
          "schemeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "inline": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InlineScheme"
              }
            ]
          }
        },
        "description": "A ceding scheme reference: either a stored scheme id or an inline definition."
      },
...
      "PensionSwitchCalcRequest": {
        "required": [
          "cedingSchemes",
          "proposedProductId",
          "retirementAge"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "cedingSchemes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CedingSchemeRef"
            }
          },
          "proposedProductId": {
            "type": "string",
            "format": "uuid"
          },
          "proposedProductChargeVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "proposedHoldings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HoldingDto"
            }
          },
          "proposedModelPortfolioId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "proposedAdviserCharges": {
            "$ref": "#/components/schemas/AdviserChargeDto"
          },
          "retirementAge": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "assumptionSetId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "overrides": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AssumptionOverrides"
              }
            ]
          },
          "redirectContributions": {
            "type": "boolean"
          },
          "clientAge": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "description": "Client age used when no client id is given (inline schemes); default 55.",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}